add getMFFloor function and change place-and-remove behavior
Place and remove behavior used to depend on a player placing certain things like internal cubes in their MF, without consideration for being inside another MF. getMFFloor added to support this.

Hopefully this fixes it, and makes it easier to string-compare if a surface is an MF surface

Changed sync area to find nearest MF first. Added findNearestMF utility function, which could be changed a bit to support more OBJs